Pull out FlexboxBoard from GameScreen
